% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult.R
\name{mult}
\alias{mult}
\title{Multiply together together two numbers}
\usage{
mult(4,5)
mult(5,10)
}
\arguments{
\item{x}{A number}

\item{y}{Another Number}
}
\value{
product The product of multiplying x and y
}
\description{
Multiply together together two numbers
}
